% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.pump}
\alias{format.pump}
\title{Format a pump pipeline}
\usage{
\method{format}{pump}(x, ...)
}
\arguments{
\item{x}{A pump object.}

\item{...}{Unused.}
}
\value{
A character vector of lines.
}
\description{
Formats a pump pipeline for display.
}
