% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{print.pump}
\alias{print.pump}
\title{Print a pump pipeline}
\usage{
\method{print}{pump}(x, ...)
}
\arguments{
\item{x}{A pump object.}

\item{...}{Unused.}
}
\description{
\code{print.pump()} displays the status of all stages in a pipeline.
}
