% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{print.pump_backend}
\alias{print.pump_backend}
\title{Print a backend}
\usage{
\method{print}{pump_backend}(x, ...)
}
\arguments{
\item{x}{A backend object.}

\item{...}{Unused.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Displays the backend's name and current worker count.
}
