% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-future.R
\name{print.pump_future_backend}
\alias{print.pump_future_backend}
\title{Print a future backend}
\usage{
\method{print}{pump_future_backend}(x, ...)
}
\arguments{
\item{x}{A future backend object.}

\item{...}{Unused.}
}
\description{
Print a future backend
}
