% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-main.R
\name{print.pump_main_backend}
\alias{print.pump_main_backend}
\title{Print a main backend}
\usage{
\method{print}{pump_main_backend}(x, ...)
}
\arguments{
\item{x}{A main backend object.}

\item{...}{Unused.}
}
\description{
Print a main backend
}
