% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mirai.R
\name{print.pump_mirai_backend}
\alias{print.pump_mirai_backend}
\title{Print a mirai backend}
\usage{
\method{print}{pump_mirai_backend}(x, ...)
}
\arguments{
\item{x}{A mirai backend object.}

\item{...}{Unused.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a mirai backend
}
