% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-parallel.R
\name{print.pump_parallel_backend}
\alias{print.pump_parallel_backend}
\title{Print a parallel backend}
\usage{
\method{print}{pump_parallel_backend}(x, ...)
}
\arguments{
\item{x}{A parallel backend object.}

\item{...}{Unused.}
}
\description{
Print a parallel backend
}
