% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{print.pump_status}
\alias{print.pump_status}
\title{Print a pump_status object}
\usage{
\method{print}{pump_status}(x, ...)
}
\arguments{
\item{x}{A pump_status object.}

\item{...}{Unused.}
}
\description{
Displays a pump_status snapshot.
}
